dbeaver/pro#6026 refactor: consolidate content type handling in useAutoContentType hook
